# Conflicts:
#	protocol.md
#	server/src/main/java/com/omegaChess/server/OCProtocol.java
#	server/src/test/java/com/omegaChess/TestOCProtocol.java

+++ server/src/test/java/com/omegaChess/TestOCServerData.java
@@ -1,18 +1,21 @@
 package com.omegaChess;
 
 import com.omegaChess.server.Invite;
+import com.omegaChess.server.Match;
 import com.omegaChess.server.OCServerData;
+import com.omegaChess.server.UserProfile;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("JUnit OCServerData Class Test")
 public class TestOCServerData {
 
     @Test
-    public void testSavingAndLoading() {
+    public void testSavingAndLoadingUserProfiles() {
+        boolean cleanup = true;
+
         // create data object to save
         OCServerData dataToSave = new OCServerData("./test-data/");
 
@@ -35,12 +38,55 @@ public class TestOCServerData {
         assertTrue(loadedData.getProfile("Daniel").getMailbox().getSent().get(0).getInvitee().equals("Falkyn"));
         assertTrue(loadedData.getProfile("Daniel").getMailbox().getReceived().get(0).getInviter().equals("Patrick"));
 
-        // cleanup
-        loadedData.deleteRootSaveFolder();
+        if (cleanup) {
+            // cleanup
+            loadedData.deleteRootSaveFolder();
 
-        // make sure cleanup worked
-        assertFalse(loadedData.rootSaveFolderExists());
+            // make sure cleanup worked
+            assertFalse(loadedData.rootSaveFolderExists());
+        }
 
+    }
+
+    @Test
+    public void testSavingAndLoadingMatches() {
+        // TODO
+    }
+
+    @Test
+    public void testSavingAndLoadingGameRecords() {
+        boolean cleanup = true;
+
+        // create data object to save
+        OCServerData dataToSave = new OCServerData("./test-data/");
+
+        // add data
+        dataToSave.createProfile("Daniel", "pass", "daniel@gmail.com");
+        dataToSave.createProfile("Falkyn", "pass", "falkyn@gmail.com");
+        Match match = new Match("Daniel", "Falkyn");
+        dataToSave.addMatch(match);
+
+        dataToSave.addToArchive(dataToSave.getMatches().get(0).endMatch("Daniel", "Falkyn", 50));
+
+        // save data
+        dataToSave.save();
+
+        // create data object to load
+        OCServerData loadedData = new OCServerData("./test-data/");
+        loadedData.load();
+
+        // make assertions to ensure loading worked properly
+        assertEquals("Falkyn", loadedData.getArchive().get(0).getWinner());
+        assertEquals("Daniel", loadedData.getArchive().get(0).getLoser());
+        assertEquals(50, loadedData.getArchive().get(0).getNumMoves());
+
+        if (cleanup) {
+            // cleanup
+            loadedData.deleteRootSaveFolder();
+
+            // make sure cleanup worked
+            assertFalse(loadedData.rootSaveFolderExists());
+        }
     }
 
 }
